Drain realtime logs from async file worker
Remove the RunLoop-based realtime log drain so realtime log records take one non-blocking hop to the file writer worker. Drop realtime records while the worker is not active to keep shutdown and pre-start behavior predictable.
